Changed Integer to Long in Repositories

+++ src/main/java/com/openrca/ecommerce/repository/ProductFeedbackRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductFeedbackRepository extends JpaRepository<ProductFeedback, Integer> {
+public interface ProductFeedbackRepository extends JpaRepository<ProductFeedback, Long> {
 
 }
