+++ src/main/java/com/openrca/ecommerce/repository/ProductSubCategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductSubCategoryRepository extends JpaRepository<ProductSubCategory,Integer> {
+public interface ProductSubCategoryRepository extends JpaRepository<ProductSubCategory,Long> {
 
 }
